+++ CHANGELOG.md
@@ -15,6 +15,7 @@ New features:
 * Calls to foreign objects with a block argument will now pass the block as the last argument.
 * `foreign.name` will now use `invokeMember` if invocable and if not use `readMember`, see `doc/contrib/interop_implicit_api.md` for details.
 * `foreign.to_f` and `foreign.to_i` will now attempt to convert to Ruby `Float` and `Integer` (#2038).
+* `foreign.equal?(other)` now uses `InteropLibrary#isIdentical(other)` and `foreign.object_id/__id__` now uses `InteropLibrary#identityHashCode()`.
 
 Bug fixes: